♻️ refactor(monitoring): extract account overview components
Move account overview cards, quota panels, token metrics, and model usage details into dedicated monitoring components.

Keep presentation helpers reusable for the table view and preserve existing page exports; validated with type-check, targeted tests, and lint.
