+++ src/features/monitoring/MonitoringCenterPage.tsx
@@ -18,11 +18,9 @@ import { Input } from '@/components/ui/Input';
 import { LoadingSpinner } from '@/components/ui/LoadingSpinner';
 import { Modal } from '@/components/ui/Modal';
 import { Select } from '@/components/ui/Select';
-import { ToggleSwitch } from '@/components/ui/ToggleSwitch';
 import { DropdownMenu, type DropdownMenuItem } from '@/components/ui/DropdownMenu';
 import {
   IconChevronDown,
-  IconChevronRight,
   IconChevronUp,
   IconChartLine,
   IconCrosshair,
@@ -30,7 +28,6 @@ import {
   IconExternalLink,
   IconFileText,
   IconInfo,
-  IconInbox,
   IconKey,
   IconMoreVertical,
   IconRefreshCw,
@@ -38,14 +35,12 @@ import {
   IconSettings,
   IconSlidersHorizontal,
   IconTimer,
-  IconTrendingUp,
 } from '@/components/ui/icons';
 import {
   buildAccountRows,
   buildApiKeyRows,
   buildRealtimeMonitorRows,
   getRangeBounds,
-  type MonitoringAccountModelSpendRow,
   type MonitoringAccountRow,
   type MonitoringApiKeyRow,
   type MonitoringCustomTimeRange,
@@ -73,13 +68,29 @@ import {
   type AccountSortState,
   type MonitoringAccountOverviewMode,
 } from '@/features/monitoring/accountOverviewState';
-import { sortAccountOverviewCardMetrics } from '@/features/monitoring/accountOverviewCardMetrics';
 import {
   buildMonitoringAccountQuotaTargetsByAccount,
   type MonitoringAccountQuotaTarget,
 } from '@/features/monitoring/accountOverviewQuotaTargets';
 import { buildRealtimeSourceDisplay } from '@/features/monitoring/realtimeSourceDisplay';
-import { MonitoringHealthStatusBar } from '@/features/monitoring/components/MonitoringHealthStatusBar';
+import {
+  AccountExpandedDetails,
+  AccountModelUsageTable,
+  AccountOverviewCard,
+  AccountStatusBadge,
+  AccountSummaryPrimary,
+  AccountTokenMetricGrid,
+} from '@/features/monitoring/components/AccountOverviewCard';
+import {
+  buildAccountSummaryMetrics,
+  formatPercent,
+  getAccountStatusTone,
+  getSuccessRateClassName,
+  type AccountQuotaEntry,
+  type AccountQuotaState,
+  type AccountQuotaWindow,
+  type AccountSummaryMetric,
+} from '@/features/monitoring/components/accountOverviewPresentation';
 import { MonitoringPanel } from '@/features/monitoring/components/MonitoringPanel';
 import { formatStatusWindowLabel } from '@/features/monitoring/model/statusWindow';
 import { useUsageData } from '@/features/monitoring/hooks/useUsageData';
@@ -102,6 +113,8 @@ import {
 import { downloadBlob } from '@/utils/download';
 import { sha256Hex } from '@/utils/apiKeyHash';
 import styles from './MonitoringCenterPage.module.scss';
+
+export { AccountExpandedDetails, AccountOverviewCard };
 
 const TIME_RANGE_OPTIONS: Array<{ value: MonitoringTimeRange; labelKey: string }> = [
   { value: 'today', labelKey: 'monitoring.range_today' },
@@ -185,42 +198,10 @@ type RealtimeLogRow = MonitoringEventRow & {
   recentPattern: boolean[];
 };
 
-type AccountQuotaWindow = {
-  id: string;
-  label: string;
-  remainingPercent: number | null;
-  resetLabel: string;
-  usageLabel: string | null;
-};
-
-type AccountQuotaEntry = {
-  key: string;
-  authLabel: string;
-  fileName: string;
-  planType: string | null;
-  windows: AccountQuotaWindow[];
-  error?: string;
-};
-
-type AccountQuotaState = {
-  status: 'idle' | 'loading' | 'success' | 'error';
-  targetKey: string;
-  entries: AccountQuotaEntry[];
-  error?: string;
-  lastRefreshedAt?: number;
-};
-
 type AccountOverviewColumn = {
   key: string;
   label: string;
   sortKey?: AccountSortKey;
-};
-
-type AccountSummaryMetric = {
-  key: string;
-  label: string;
-  value: string;
-  valueClassName?: string;
 };
 
 type PaginationState<T> = {
@@ -243,8 +224,6 @@ type PaginationControlsProps = {
   onPageSizeChange: (pageSize: number) => void;
   t: TFunction;
 };
-
-const formatPercent = (value: number) => `${(value * 100).toFixed(1)}%`;
 
 const ensureSelectedOption = <T extends { value: string; label: string }>(
   options: T[],
@@ -318,101 +297,12 @@ const buildRealtimeMetaText = (row: MonitoringEventRow) => {
   return maskSensitiveText(text || '-');
 };
 
-const PREMIUM_CODEX_PLAN_TYPES = new Set(['pro', 'prolite', 'pro-lite', 'pro_lite']);
-
-const getCodexPlanLabel = (planType: string | null | undefined, t: TFunction): string | null => {
-  const normalized = normalizePlanType(planType);
-  if (!normalized) return null;
-  if (normalized === 'pro') return t('codex_quota.plan_pro');
-  if (PREMIUM_CODEX_PLAN_TYPES.has(normalized) && normalized !== 'pro') {
-    return t('codex_quota.plan_prolite');
-  }
-  if (normalized === 'plus') return t('codex_quota.plan_plus');
-  if (normalized === 'team') return t('codex_quota.plan_team');
-  if (normalized === 'free') return t('codex_quota.plan_free');
-  return planType || normalized;
-};
-
-const buildAccountSecondaryText = (row: MonitoringAccountRow) => {
-  const primaryText = row.displayAccount || row.account;
-  if (row.account && row.account !== primaryText) {
-    return row.account;
-  }
-
-  const extraAuthLabels = row.authLabels.filter((label) => label && label !== primaryText);
-  if (extraAuthLabels.length > 0) {
-    return joinShort(extraAuthLabels, 2);
-  }
-  const extraChannels = row.channels.filter(
-    (label) => label && label !== '-' && label !== primaryText
-  );
-  if (extraChannels.length > 0) {
-    return joinShort(extraChannels, 2);
-  }
-  return '';
-};
-
 const buildAccountOptionLabel = (row: MonitoringAccountRow) => {
   if (!row.displayAccount || row.displayAccount === row.account) {
     return row.account;
   }
   return `${row.displayAccount} / ${row.account}`;
 };
-
-const buildAccountSummaryMetrics = (
-  row: MonitoringAccountRow,
-  hasPrices: boolean,
-  locale: string,
-  t: TFunction
-): AccountSummaryMetric[] => [
-  {
-    key: 'total-calls',
-    label: t('monitoring.total_calls'),
-    value: formatCompactNumber(row.totalCalls),
-  },
-  {
-    key: 'success-calls',
-    label: t('monitoring.success_calls'),
-    value: formatCompactNumber(row.successCalls),
-    valueClassName: styles.goodText,
-  },
-  {
-    key: 'failure-calls',
-    label: t('monitoring.failure_calls'),
-    value: formatCompactNumber(row.failureCalls),
-    valueClassName: row.failureCalls > 0 ? styles.badText : undefined,
-  },
-  {
-    key: 'total-tokens',
-    label: t('monitoring.total_tokens'),
-    value: formatCompactNumber(row.totalTokens),
-  },
-  {
-    key: 'input-tokens',
-    label: t('monitoring.input_tokens'),
-    value: formatCompactNumber(row.inputTokens),
-  },
-  {
-    key: 'output-tokens',
-    label: t('monitoring.output_tokens'),
-    value: formatCompactNumber(row.outputTokens),
-  },
-  {
-    key: 'cached-tokens',
-    label: t('monitoring.cached_tokens'),
-    value: formatCompactNumber(row.cachedTokens),
-  },
-  {
-    key: 'estimated-cost',
-    label: t('monitoring.estimated_cost'),
-    value: hasPrices ? formatUsd(row.totalCost) : '--',
-  },
-  {
-    key: 'latest-request-time',
-    label: t('monitoring.latest_request_time'),
-    value: new Date(row.lastSeenAt).toLocaleString(locale),
-  },
-];
 
 const buildApiKeySecondaryText = (row: MonitoringApiKeyRow) => {
   if (row.isUnknown) {
@@ -713,663 +603,6 @@ const EMPTY_ACCOUNT_AUTH_STATE: MonitoringAccountAuthState = {
   toggleableFileNames: [],
   enabledState: 'unavailable',
 };
-
-const getAccountStatusTone = (authState: MonitoringAccountAuthState) => {
-  switch (authState.enabledState) {
-    case 'enabled':
-      return 'enabled';
-    case 'disabled':
-      return 'disabled';
-    case 'mixed':
-      return 'mixed';
-    case 'unavailable':
-    default:
-      return 'unavailable';
-  }
-};
-
-const getAccountStatusLabel = (authState: MonitoringAccountAuthState, t: TFunction) => {
-  switch (authState.enabledState) {
-    case 'enabled':
-      return t('monitoring.account_overview_enabled_state_enabled');
-    case 'disabled':
-      return t('monitoring.account_overview_enabled_state_disabled');
-    case 'mixed':
-      return t('monitoring.account_overview_enabled_state_mixed');
-    case 'unavailable':
-    default:
-      return t('monitoring.account_overview_enabled_state_unavailable');
-  }
-};
-
-const getAccountStatusDotClassName = (tone: string) => {
-  switch (tone) {
-    case 'enabled':
-      return styles.accountStatusDotEnabled;
-    case 'disabled':
-      return styles.accountStatusDotDisabled;
-    case 'mixed':
-      return styles.accountStatusDotMixed;
-    case 'unavailable':
-    default:
-      return styles.accountStatusDotUnavailable;
-  }
-};
-
-const getSuccessRateClassName = (rate: number) =>
-  rate >= 0.95 ? styles.goodText : rate >= 0.85 ? styles.warnText : styles.badText;
-
-function AccountStatusBadge({
-  authState,
-  t,
-}: {
-  authState: MonitoringAccountAuthState;
-  t: TFunction;
-}) {
-  const tone = getAccountStatusTone(authState);
-  const label = getAccountStatusLabel(authState, t);
-
-  return (
-    <span
-      className={[styles.accountStatusBadge, styles[`accountStatusBadge${tone}`]]
-        .filter(Boolean)
-        .join(' ')}
-      title={label}
-    >
-      <span
-        className={[styles.accountStatusDot, getAccountStatusDotClassName(tone)]
-          .filter(Boolean)
-          .join(' ')}
-        aria-hidden="true"
-      />
-      {label}
-    </span>
-  );
-}
-
-function AccountSummaryPrimary({
-  row,
-  expanded,
-  onToggle,
-  statusTone = 'enabled',
-  showSecondary = true,
-}: {
-  row: MonitoringAccountRow;
-  expanded: boolean;
-  onToggle: () => void;
-  statusTone?: string;
-  showSecondary?: boolean;
-}) {
-  const secondaryText = buildAccountSecondaryText(row);
-  const accountLabel = row.displayAccount || row.account;
-
-  return (
-    <button
-      type="button"
-      className={[
-        styles.accountButton,
-        expanded ? styles.expandedAccountButton : '',
-        statusTone === 'disabled' || statusTone === 'unavailable' ? styles.accountButtonMuted : '',
-      ]
-        .filter(Boolean)
-        .join(' ')}
-      onClick={onToggle}
-      aria-expanded={expanded}
-      title={accountLabel}
-    >
-      <span className={styles.accountExpandGlyph} aria-hidden="true">
-        {expanded ? <IconChevronUp size={15} /> : <IconChevronDown size={15} />}
-      </span>
-      <span className={styles.accountIdentityLine}>
-        <span
-          className={[styles.accountStatusDot, getAccountStatusDotClassName(statusTone)]
-            .filter(Boolean)
-            .join(' ')}
-          aria-hidden="true"
-        />
-        <span className={styles.accountButtonLabel}>{accountLabel}</span>
-      </span>
-      {showSecondary && secondaryText ? <small>{secondaryText}</small> : null}
-    </button>
-  );
-}
-
-function AccountQuotaPanel({
-  quotaState,
-  locale,
-  t,
-  onRefreshQuota,
-}: {
-  quotaState?: AccountQuotaState;
-  locale: string;
-  t: TFunction;
-  onRefreshQuota: () => void;
-}) {
-  const quotaEntries = quotaState?.entries ?? [];
-  const quotaLoading = quotaState?.status === 'loading';
-  const lastQuotaSync =
-    quotaState?.lastRefreshedAt && Number.isFinite(quotaState.lastRefreshedAt)
-      ? new Date(quotaState.lastRefreshedAt).toLocaleString(locale)
-      : '';
-  const singleQuotaEntry = quotaEntries.length === 1 ? quotaEntries[0] : null;
-  const singlePlanLabel = singleQuotaEntry ? getCodexPlanLabel(singleQuotaEntry.planType, t) : null;
-  const quotaMetaText = [
-    singlePlanLabel ? `${t('codex_quota.plan_label')}: ${singlePlanLabel}` : '',
-    lastQuotaSync ? `${t('monitoring.last_sync')}: ${lastQuotaSync}` : '',
-  ]
-    .filter(Boolean)
-    .join(' · ');
-
-  const renderQuotaWindows = (windows: AccountQuotaWindow[]) => (
-    <div className={styles.quotaWindowList}>
-      {windows.map((window) => {
-        const percentLabel =
-          window.remainingPercent === null ? '--' : `${Math.round(window.remainingPercent)}%`;
-        const barStyle =
-          window.remainingPercent === null
-            ? undefined
-            : { width: `${Math.max(0, Math.min(100, window.remainingPercent))}%` };
-
-        return (
-          <div key={window.id} className={styles.quotaWindowRow}>
-            <div className={styles.quotaWindowHeader}>
-              <span>{window.label}</span>
-              <strong>{percentLabel}</strong>
-            </div>
-            <div className={styles.quotaProgressTrack}>
-              <span className={styles.quotaProgressBar} style={barStyle} />
-            </div>
-            <div className={styles.quotaWindowMeta}>
-              <small>{`${t('monitoring.account_quota_reset_at')}: ${window.resetLabel}`}</small>
-              {window.usageLabel ? <small>{window.usageLabel}</small> : null}
-            </div>
-          </div>
-        );
-      })}
-    </div>
-  );
-
-  const renderRefreshButton = () => (
-    <button
-      type="button"
-      className={styles.quotaRefreshButton}
-      onClick={onRefreshQuota}
-      disabled={quotaLoading}
-    >
-      <IconRefreshCw
-        size={14}
-        className={quotaLoading ? styles.refreshIconSpinning : styles.refreshIcon}
-      />
-      <span>{t('codex_quota.refresh_button')}</span>
-    </button>
-  );
-
-  const renderStateMessage = (message: ReactNode, hint?: ReactNode, retry = false) => (
-    <div className={styles.quotaStateMessage}>
-      <span>{message}</span>
-      {hint ? <small>{hint}</small> : null}
-      {retry ? (
-        <button
-          type="button"
-          className={styles.quotaRetryButton}
-          onClick={onRefreshQuota}
-          disabled={quotaLoading}
-        >
-          <IconRefreshCw
-            size={14}
-            className={quotaLoading ? styles.refreshIconSpinning : styles.refreshIcon}
-          />
-          <span>{t('codex_quota.retry_button')}</span>
-        </button>
-      ) : null}
-    </div>
-  );
-
-  return (
-    <section className={styles.quotaSection}>
-      <div className={styles.quotaSectionHeader}>
-        <div className={styles.quotaSectionTitleGroup}>
-          <strong>{t('codex_quota.title')}</strong>
-          {quotaMetaText ? <span>{quotaMetaText}</span> : null}
-        </div>
-        {renderRefreshButton()}
-      </div>
-
-      {quotaLoading && quotaEntries.length === 0
-        ? renderStateMessage(t('codex_quota.loading'))
-        : null}
-
-      {!quotaLoading && quotaState?.status === 'error' && quotaEntries.length === 0
-        ? renderStateMessage(
-            t('codex_quota.load_failed', {
-              message: quotaState.error || t('common.unknown_error'),
-            }),
-            undefined,
-            true
-          )
-        : null}
-
-      {!quotaLoading && quotaState?.status === 'success' && quotaEntries.length === 0
-        ? renderStateMessage(t('codex_quota.empty_windows'), t('codex_quota.idle'))
-        : null}
-
-      {!quotaState && quotaEntries.length === 0
-        ? renderStateMessage(t('codex_quota.empty_windows'), t('codex_quota.idle'))
-        : null}
-
-      {singleQuotaEntry ? (
-        singleQuotaEntry.error ? (
-          renderStateMessage(
-            t('codex_quota.load_failed', { message: singleQuotaEntry.error }),
-            undefined,
-            true
-          )
-        ) : singleQuotaEntry.windows.length > 0 ? (
-          renderQuotaWindows(singleQuotaEntry.windows)
-        ) : (
-          renderStateMessage(t('codex_quota.empty_windows'), t('codex_quota.idle'))
-        )
-      ) : quotaEntries.length > 0 ? (
-        <div className={styles.quotaEntryGrid}>
-          {quotaEntries.map((entry) => {
-            const planLabel = getCodexPlanLabel(entry.planType, t);
-            return (
-              <div key={entry.key} className={styles.quotaEntryCard}>
-                <div className={styles.quotaEntryHeader}>
-                  <div className={styles.quotaEntryMain}>
-                    <strong>{entry.authLabel}</strong>
-                    <small>
-                      {planLabel ? `${t('codex_quota.plan_label')}: ${planLabel}` : entry.fileName}
-                    </small>
-                  </div>
-                </div>
-
-                {entry.error
-                  ? renderStateMessage(
-                      t('codex_quota.load_failed', { message: entry.error }),
-                      undefined,
-                      true
-                    )
-                  : entry.windows.length > 0
-                    ? renderQuotaWindows(entry.windows)
-                    : renderStateMessage(t('codex_quota.empty_windows'), t('codex_quota.idle'))}
-              </div>
-            );
-          })}
-        </div>
-      ) : null}
-    </section>
-  );
-}
-
-function AccountTokenMetricGrid({
-  metrics,
-  t,
-  variant = 'card',
-}: {
-  metrics: AccountSummaryMetric[];
-  t: TFunction;
-  variant?: 'card' | 'table';
-}) {
-  const getTokenMetricIcon = (key: string) => {
-    if (key === 'input-tokens') return <IconInbox size={13} />;
-    if (key === 'output-tokens') return <IconTrendingUp size={13} />;
-    if (key === 'cached-tokens') return <IconTimer size={13} />;
-    return <IconChartLine size={13} />;
-  };
-  const getTokenMetricToneClassName = (key: string) => {
-    if (key === 'input-tokens') return styles.accountMetricIconInput;
-    if (key === 'output-tokens') return styles.accountMetricIconOutput;
-    if (key === 'cached-tokens') return styles.accountMetricIconCached;
-    return styles.accountMetricIconTotal;
-  };
-
-  if (variant === 'table') {
-    const tokenStructureMetrics = metrics.filter((metric) =>
-      ['input-tokens', 'output-tokens', 'cached-tokens'].includes(metric.key)
-    );
-    const getTokenStructureRowToneClassName = (key: string) => {
-      if (key === 'input-tokens') return styles.tokenStructureRowInput;
-      if (key === 'output-tokens') return styles.tokenStructureRowOutput;
-      if (key === 'cached-tokens') return styles.tokenStructureRowCached;
-      return '';
-    };
-
-    return (
-      <section className={styles.accountTokenStructurePanel}>
-        <div className={styles.accountSectionHeader}>
-          <strong>{t('monitoring.account_overview_token_structure')}</strong>
-        </div>
-        <div className={styles.tokenStructureRowList}>
-          {tokenStructureMetrics.map((metric) => (
-            <div
-              key={metric.key}
-              className={[styles.tokenStructureRow, getTokenStructureRowToneClassName(metric.key)]
-                .filter(Boolean)
-                .join(' ')}
-            >
-              <span className={styles.tokenStructureRowLeft}>
-                <span className={styles.tokenStructureRowIcon} aria-hidden="true">
-                  {getTokenMetricIcon(metric.key)}
-                </span>
-                <span className={styles.tokenStructureRowLabel}>{metric.label}</span>
-              </span>
-              <strong
-                className={[styles.tokenStructureRowValue, metric.valueClassName]
-                  .filter(Boolean)
-                  .join(' ')}
-              >
-                {metric.value}
-              </strong>
-            </div>
-          ))}
-        </div>
-      </section>
-    );
-  }
-
-  return (
-    <section className={styles.accountTokenPanel}>
-      <div className={styles.accountSectionHeader}>
-        <strong>{t('monitoring.account_overview_tokens_title')}</strong>
-      </div>
-      <div className={styles.accountOverviewMetricGrid}>
-        {metrics.map((metric) => (
-          <div key={metric.key} className={styles.accountOverviewMetricCard}>
-            <span className={styles.accountOverviewMetricLabel}>
-              <span
-                className={[styles.accountMetricIcon, getTokenMetricToneClassName(metric.key)]
-                  .filter(Boolean)
-                  .join(' ')}
-                aria-hidden="true"
-              >
-                {getTokenMetricIcon(metric.key)}
-              </span>
-              {metric.label}
-            </span>
-            <strong
-              className={[styles.accountOverviewMetricValue, metric.valueClassName]
-                .filter(Boolean)
-                .join(' ')}
-            >
-              {metric.value}
-            </strong>
-          </div>
-        ))}
-      </div>
-    </section>
-  );
-}
-
-function AccountHealthStatusPanel({
-  row,
-  hasPrices,
-  locale,
-  t,
-  statusData,
-  scopeText,
-}: {
-  row: MonitoringAccountRow;
-  hasPrices: boolean;
-  locale: string;
-  t: TFunction;
-  statusData: StatusBarData;
-  scopeText: string;
-}) {
-  const healthMetrics = [
-    {
-      key: 'total-calls',
-      label: t('monitoring.total_calls'),
-      value: formatCompactNumber(row.totalCalls),
-    },
-    {
-      key: 'success-calls',
-      label: t('stats.success'),
-      value: formatCompactNumber(row.successCalls),
-      className: styles.goodText,
-    },
-    {
-      key: 'failure-calls',
-      label: t('stats.failure'),
-      value: formatCompactNumber(row.failureCalls),
-      className: row.failureCalls > 0 ? styles.badText : undefined,
-    },
-    {
-      key: 'estimated-cost',
-      label: t('monitoring.estimated_cost'),
-      value: hasPrices ? formatUsd(row.totalCost) : '--',
-      className: styles.primaryText,
-    },
-    {
-      key: 'success-rate',
-      label: t('monitoring.column_success_rate'),
-      value: formatPercent(row.successRate),
-      className: getSuccessRateClassName(row.successRate),
-    },
-  ];
-
-  return (
-    <section className={styles.accountOverviewStatusSection}>
-      <div className={styles.accountSectionHeader}>
-        <strong>{t('monitoring.account_overview_health_label')}</strong>
-        <span
-          className={styles.accountSectionInfo}
-          title={t('monitoring.account_overview_health_hint')}
-        >
-          <IconInfo size={14} />
-        </span>
-      </div>
-      <div className={styles.healthMetricGrid}>
-        {healthMetrics.map((metric) => (
-          <div key={metric.key} className={styles.healthMetricItem}>
-            <span>{metric.label}</span>
-            <strong className={metric.className}>{metric.value}</strong>
-          </div>
-        ))}
-      </div>
-      <MonitoringHealthStatusBar statusData={statusData} locale={locale} t={t} showRate={false} />
-      <div className={styles.accountScopeText}>{scopeText}</div>
-    </section>
-  );
-}
-
-function AccountModelUsageList({
-  row,
-  hasPrices,
-  locale,
-  t,
-  limit = 2,
-}: {
-  row: { id: string; models: MonitoringAccountModelSpendRow[] };
-  hasPrices: boolean;
-  locale: string;
-  t: TFunction;
-  limit?: number;
-}) {
-  const [showAll, setShowAll] = useState(false);
-  const [expandedModels, setExpandedModels] = useState<Record<string, boolean>>({});
-  const hasExtraModels = row.models.length > limit;
-  const visibleModels = showAll ? row.models : row.models.slice(0, limit);
-  const toggleModel = (key: string) =>
-    setExpandedModels((previous) => ({ ...previous, [key]: !previous[key] }));
-
-  return (
-    <section className={styles.accountModelListPanel}>
-      <div className={styles.accountSectionHeader}>
-        <strong>
-          {t('monitoring.account_overview_models_top', {
-            count: Math.min(limit, row.models.length || limit),
-          })}
-        </strong>
-        {hasExtraModels ? (
-          <button
-            type="button"
-            className={styles.accountModelViewAllButton}
-            onClick={() => setShowAll((previous) => !previous)}
-          >
-            {showAll
-              ? t('monitoring.account_overview_collapse_models')
-              : t('monitoring.account_overview_view_all')}
-          </button>
-        ) : null}
-      </div>
-
-      {visibleModels.length > 0 ? (
-        <div className={styles.accountModelList}>
-          {visibleModels.map((model) => {
-            const modelKey = `${row.id}-${model.model}`;
-            const isModelExpanded = Boolean(expandedModels[modelKey]);
-            return (
-              <div key={modelKey} className={styles.accountModelItem}>
-                <button
-                  type="button"
-                  className={styles.accountModelRow}
-                  onClick={() => toggleModel(modelKey)}
-                  aria-expanded={isModelExpanded}
-                >
-                  <span className={styles.accountModelName} title={model.model}>
-                    {model.model}
-                  </span>
-                  <span className={styles.accountModelMetaLine}>
-                    <span className={styles.accountModelStat}>
-                      <small>{t('monitoring.account_overview_model_calls_short')}</small>
-                      <strong>{formatCompactNumber(model.totalCalls)}</strong>
-                    </span>
-                    <span className={styles.accountModelStat}>
-                      <small>{t('monitoring.account_overview_model_success_rate_short')}</small>
-                      <strong className={getSuccessRateClassName(model.successRate)}>
-                        {formatPercent(model.successRate)}
-                      </strong>
-                    </span>
-                    <span className={styles.accountModelStat}>
-                      <small>{t('monitoring.account_overview_model_total_tokens_short')}</small>
-                      <strong>{formatCompactNumber(model.totalTokens)}</strong>
-                    </span>
-                    <span className={styles.accountModelStat}>
-                      <small>{t('monitoring.account_overview_model_total_cost_short')}</small>
-                      <strong>{hasPrices ? formatUsd(model.totalCost) : '--'}</strong>
-                    </span>
-                    <span className={styles.accountModelChevron} aria-hidden="true">
-                      {isModelExpanded ? (
-                        <IconChevronDown size={14} />
-                      ) : (
-                        <IconChevronRight size={14} />
-                      )}
-                    </span>
-                  </span>
-                </button>
-                {isModelExpanded ? (
-                  <div className={styles.accountModelExpanded}>
-                    <div className={styles.accountModelExpandedItem}>
-                      <small>{t('monitoring.input_tokens')}</small>
-                      <strong>{formatCompactNumber(model.inputTokens)}</strong>
-                    </div>
-                    <div className={styles.accountModelExpandedItem}>
-                      <small>{t('monitoring.output_tokens')}</small>
-                      <strong>{formatCompactNumber(model.outputTokens)}</strong>
-                    </div>
-                    <div className={styles.accountModelExpandedItem}>
-                      <small>{t('monitoring.cached_tokens')}</small>
-                      <strong>{formatCompactNumber(model.cachedTokens)}</strong>
-                    </div>
-                    <div className={styles.accountModelExpandedItem}>
-                      <small>{t('monitoring.latest_request_time')}</small>
-                      <strong>{new Date(model.lastSeenAt).toLocaleString(locale)}</strong>
-                    </div>
-                  </div>
-                ) : null}
-              </div>
-            );
-          })}
-        </div>
-      ) : (
-        <div className={styles.emptyBlockSmall}>{t('monitoring.account_overview_no_models')}</div>
-      )}
-    </section>
-  );
-}
-
-function AccountModelUsageTable({
-  row,
-  hasPrices,
-  locale,
-  t,
-  limit = 2,
-}: {
-  row: { id: string; models: MonitoringAccountModelSpendRow[] };
-  hasPrices: boolean;
-  locale: string;
-  t: TFunction;
-  limit?: number;
-}) {
-  const [showAll, setShowAll] = useState(false);
-  const hasExtraModels = row.models.length > limit;
-  const visibleModels = showAll ? row.models : row.models.slice(0, limit);
-  const modelCountForTitle = Math.min(limit, row.models.length || limit);
-
-  return (
-    <section className={styles.accountModelTablePanel}>
-      <div className={styles.accountSectionHeader}>
-        <strong>
-          {t('monitoring.account_overview_models_top', {
-            count: modelCountForTitle,
-          })}
-        </strong>
-        <button
-          type="button"
-          className={styles.accountModelViewAllButton}
-          onClick={() => setShowAll((previous) => !previous)}
-          disabled={!hasExtraModels}
-        >
-          {showAll
-            ? t('monitoring.account_overview_collapse_models')
-            : t('monitoring.account_overview_view_all')}
-        </button>
-      </div>
-      {visibleModels.length > 0 ? (
-        <table className={styles.accountModelTable}>
-          <thead>
-            <tr>
-              <th>{t('usage_stats.model_price_model')}</th>
-              <th>{t('monitoring.account_overview_model_calls_short')}</th>
-              <th>{t('monitoring.account_overview_model_success_rate_short')}</th>
-              <th>{t('monitoring.account_overview_model_input_tokens_short')}</th>
-              <th>{t('monitoring.account_overview_model_output_tokens_short')}</th>
-              <th>{t('monitoring.account_overview_model_cached_tokens_short')}</th>
-              <th>{t('monitoring.account_overview_model_total_tokens_short')}</th>
-              <th>{t('monitoring.account_overview_model_total_cost_short')}</th>
-              <th>{t('monitoring.latest_request_time')}</th>
-            </tr>
-          </thead>
-          <tbody>
-            {visibleModels.map((model) => (
-              <tr key={`${row.id}-${model.model}`}>
-                <td>
-                  <span className={styles.accountModelName} title={model.model}>
-                    {model.model}
-                  </span>
-                </td>
-                <td>{formatCompactNumber(model.totalCalls)}</td>
-                <td className={getSuccessRateClassName(model.successRate)}>
-                  {formatPercent(model.successRate)}
-                </td>
-                <td>{formatCompactNumber(model.inputTokens)}</td>
-                <td>{formatCompactNumber(model.outputTokens)}</td>
-                <td>{formatCompactNumber(model.cachedTokens)}</td>
-                <td>{formatCompactNumber(model.totalTokens)}</td>
-                <td>{hasPrices ? formatUsd(model.totalCost) : '--'}</td>
-                <td>{new Date(model.lastSeenAt).toLocaleString(locale)}</td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
-      ) : (
-        <div className={styles.emptyBlockSmall}>{t('monitoring.account_overview_no_models')}</div>
-      )}
-    </section>
-  );
-}
-
 function ApiKeySummaryPrimary({
   row,
   expanded,
@@ -1432,202 +665,6 @@ function ApiKeyExpandedDetails({
     </div>
   );
 }
-
-export function AccountExpandedDetails({
-  row,
-  hasPrices,
-  locale,
-  t,
-  summaryMetrics,
-  quotaState,
-  onRefreshQuota,
-  variant,
-}: {
-  row: MonitoringAccountRow;
-  hasPrices: boolean;
-  locale: string;
-  t: TFunction;
-  summaryMetrics: AccountSummaryMetric[];
-  quotaState?: AccountQuotaState;
-  onRefreshQuota: () => void;
-  variant: 'card' | 'table';
-}) {
-  const tokenMetrics = sortAccountOverviewCardMetrics(summaryMetrics);
-
-  if (variant === 'table') {
-    return (
-      <div className={styles.expandedAccountDetails}>
-        <AccountQuotaPanel
-          quotaState={quotaState}
-          locale={locale}
-          t={t}
-          onRefreshQuota={onRefreshQuota}
-        />
-        <div className={styles.accountStructureModelPanel}>
-          <AccountTokenMetricGrid metrics={tokenMetrics} t={t} variant="table" />
-          <AccountModelUsageTable row={row} hasPrices={hasPrices} locale={locale} t={t} />
-        </div>
-      </div>
-    );
-  }
-
-  return (
-    <div className={styles.accountOverviewCardBody}>
-      <AccountQuotaPanel
-        quotaState={quotaState}
-        locale={locale}
-        t={t}
-        onRefreshQuota={onRefreshQuota}
-      />
-      <AccountModelUsageList row={row} hasPrices={hasPrices} locale={locale} t={t} />
-    </div>
-  );
-}
-
-export function AccountOverviewCard({
-  row,
-  authState,
-  hasPrices,
-  locale,
-  t,
-  isExpanded,
-  isFocused,
-  statusData,
-  scopeText,
-  quotaState,
-  statusUpdating,
-  onToggle,
-  onFocus,
-  onToggleEnabled,
-  onRefreshQuota,
-}: {
-  row: MonitoringAccountRow;
-  authState: MonitoringAccountAuthState;
-  hasPrices: boolean;
-  locale: string;
-  t: TFunction;
-  isExpanded: boolean;
-  isFocused: boolean;
-  statusData: StatusBarData;
-  scopeText: string;
-  quotaState?: AccountQuotaState;
-  statusUpdating: boolean;
-  onToggle: () => void;
-  onFocus: () => void;
-  onToggleEnabled: (enabled: boolean) => void;
-  onRefreshQuota: () => void;
-}) {
-  const summaryMetrics = buildAccountSummaryMetrics(row, hasPrices, locale, t);
-  const cardMetrics = sortAccountOverviewCardMetrics(summaryMetrics);
-  const canToggleEnabled = authState.enabledState !== 'unavailable';
-  const toggleChecked = authState.enabledState === 'enabled';
-  const statusTone = getAccountStatusTone(authState);
-  const secondaryText = buildAccountSecondaryText(row);
-  const latestRequestText = new Date(row.lastSeenAt).toLocaleString(locale);
-
-  return (
-    <Card
-      className={[
-        styles.accountOverviewCard,
-        isExpanded ? styles.accountOverviewCardExpanded : '',
-        isFocused ? styles.accountOverviewCardFocused : '',
-        authState.enabledState === 'disabled' ? styles.accountOverviewCardDisabled : '',
-      ]
-        .filter(Boolean)
-        .join(' ')}
-    >
-      <div className={styles.accountOverviewCardHeader}>
-        <div className={styles.accountTitleRow}>
-          <AccountSummaryPrimary
-            row={row}
-            expanded={isExpanded}
-            onToggle={onToggle}
-            statusTone={statusTone}
-            showSecondary={false}
-          />
-          <div className={styles.accountEnabledControl}>
-            <span className={styles.accountEnabledLabel}>
-              {t('monitoring.account_overview_enabled_label_short')}
-            </span>
-            {authState.enabledState === 'mixed' ? (
-              <div className={styles.accountOverviewToggleActions}>
-                <button
-                  type="button"
-                  className={styles.inlineActionButton}
-                  onClick={() => onToggleEnabled(true)}
-                  disabled={statusUpdating}
-                >
-                  {t('monitoring.account_overview_enable_all')}
-                </button>
-                <button
-                  type="button"
-                  className={styles.inlineActionButton}
-                  onClick={() => onToggleEnabled(false)}
-                  disabled={statusUpdating}
-                >
-                  {t('monitoring.account_overview_disable_all')}
-                </button>
-              </div>
-            ) : (
-              <ToggleSwitch
-                ariaLabel={t('monitoring.account_overview_enabled_label')}
-                checked={toggleChecked}
-                disabled={!canToggleEnabled || statusUpdating}
-                onChange={onToggleEnabled}
-              />
-            )}
-          </div>
-        </div>
-        <div className={styles.accountMetaRow}>
-          {secondaryText ? (
-            <span className={styles.accountOverviewCardTimestamp} title={secondaryText}>
-              {secondaryText}
-            </span>
-          ) : null}
-          {secondaryText ? <span className={styles.accountMetaSeparator}>·</span> : null}
-          <span className={styles.accountOverviewCardTimestamp}>
-            {`${t('monitoring.latest_request_time')}: ${latestRequestText}`}
-          </span>
-          <button
-            type="button"
-            className={`${styles.inlineActionButton} ${styles.accountFocusButton}`}
-            onClick={onFocus}
-          >
-            <IconCrosshair size={12} aria-hidden="true" />
-            <span>
-              {isFocused ? t('monitoring.restore_account_scope') : t('monitoring.focus_account')}
-            </span>
-          </button>
-        </div>
-      </div>
-
-      <AccountHealthStatusPanel
-        row={row}
-        hasPrices={hasPrices}
-        locale={locale}
-        t={t}
-        statusData={statusData}
-        scopeText={scopeText}
-      />
-
-      <AccountTokenMetricGrid metrics={cardMetrics} t={t} />
-
-      {isExpanded ? (
-        <AccountExpandedDetails
-          row={row}
-          hasPrices={hasPrices}
-          locale={locale}
-          t={t}
-          summaryMetrics={summaryMetrics}
-          quotaState={quotaState}
-          onRefreshQuota={onRefreshQuota}
-          variant="card"
-        />
-      ) : null}
-    </Card>
-  );
-}
-
 export function MonitoringCenterPage() {
   const { t, i18n } = useTranslation();
   const config = useConfigStore((state) => state.config);
